requirements.txt
PyQt6
PyQt6-WebEngine
openai
python-dotenv
pytest 
